requirements.txt
streamlit==1.32.0
firebase-admin==6.4.0
pyrebase4==4.7.1
folium==0.15.1
streamlit-folium==0.15.1
pytest==8.0.0
python-dotenv==1.0.1
